+++ packages/common/src/hooks/index.ts
@@ -34,4 +34,5 @@ export * from './useChangePasswordFormConfiguration'
 export * from './useChangeEmailFormConfiguration'
 export * from './useHarmonyField'
 export * from './useAudioRewards'
-export * from './useMetadata'
+export * from './useCollectionMetadata'
+export * from './useTrackMetadata'
